+++ client/src/components/MessageList/useMessageList.ts
@@ -1,0 +1,19 @@
+import useSWR from "swr";
+
+const fetcher = (url: string) =>
+  fetch(url, {
+    headers: {
+      Accept: "application/json",
+    },
+  })
+    .then((res) => res.json())
+    .then((res) => res.results);
+
+export const useMessageList = () => {
+  const { data: ltds, error } = useSWR(
+    "https://icanhazdadjoke.com/search?page=2",
+    fetcher
+  );
+
+  return { ltds, error };
+};
